sizeable refactor of code layout
